allow InetSource to assign data to 'blob', URL to 'rstring' attribute
The first change allows the InetSource operator to assign data received
from URLs to an output attribute of type 'blob'. This allows XML data to
be passed directly to the XMLParse operator, and allows JPEG data to be
passed directly to OpenCV-based operators.

The second change allows the InetSource operator to assign the URL the
data to an optional second output attribute of type 'rstring'.